Fixed searching in messages
Discovered that Azure Table Storage doesn't have the function to partially search in fields that are not the partition key or row key. So from what I found, the best way to go around this is retrieving all the messages into the memory and searching in the application. Or we need to choose another database to store messages.